Updated from u16 to f32.

+++ include/label.hpp
@@ -8,7 +8,7 @@ namespace SUI {
 namespace Widget {
     class Label : public IRenderable, public ITextArea {
         public:
-            Label(const char *text, u16 x, u16 y, u16 padding=0);
+            Label(const char *text, f32 x, f32 y, u16 padding=0);
             ~Label();
 
             virtual void Render(void);
